Phase 4g: extract NervesHub.Devices.Updates sub-context
Extracts the device update-orchestration engine out of the Devices god-module
into a new NervesHub.Devices.Updates module. Built on Phase 4f.

Moved (deciding whether/what/when to update a device, plus update settings):
firmware_validated, available_for_update, available_for_priority_update,
resolve_update, verify_update_eligibility, update_attempted,
failure_threshold_met?, failure_rate_met?, device_in_penalty_box?,
device_matches_deployment_group?, enable_updates, disable_updates,
toggle_automatic_updates, clear_penalty_box, update_blocked_until, plus their
private query helpers (build_available_devices_query and friends,
put_device_in_penalty_box, updates_blocked?). They call back into Devices for
core ops (get_device/update_device/update_device_with_audit).

devices.ex drops from ~1537 to ~984 lines; several now-unused aliases removed.
Adds a .dialyzer_ignore.exs entry for the new module (update_attempted uses an
Ecto.Multi chain).

Call sites repointed to Updates across lib and test. Notably the dynamic
dispatch in Devices.BulkActions — `Task.Supervisor.async(Tasks, <mod>, :fn, ..)`
and `apply(<mod>, fun, ..)` — now carries the target module (Devices for
tag/move, Updates for enable/disable/clear), and the orchestrator test's Mimic
wiring (`allow(Updates, self(), pid)` + `Updates |> expect(:available_for_update)`)
follows the moved function across the orchestrator's separate process.

Verification: mix format (bare, exit 0), credo, dialyzer (0 unskipped), full
test suite (1396 passing).
