+++ lib/nerves_hub_web/live/devices/show.ex
@@ -6,6 +6,7 @@ defmodule NervesHubWeb.Live.Devices.Show do
   alias NervesHub.Devices
   alias NervesHub.Devices.Connections
   alias NervesHub.Devices.Pinning
+  alias NervesHub.Devices.Updates
   alias NervesHub.Extensions.Health
   alias NervesHub.FirmwareUpdates
   alias NervesHub.Tracker
@@ -108,7 +109,7 @@ defmodule NervesHubWeb.Live.Devices.Show do
     |> assign(:device_connection, device.latest_connection)
     |> load_inprogress_firmware_update()
     |> async_console_status_check()
-    |> assign(:update_information, Devices.resolve_update(device))
+    |> assign(:update_information, Updates.resolve_update(device))
     |> then(fn socket ->
       if(payload.status == "online", do: clear_flash(socket), else: socket)
     end)
@@ -261,7 +262,7 @@ defmodule NervesHubWeb.Live.Devices.Show do
 
     authorized!(:"device:clear-penalty-box", current_scope)
 
-    {:ok, updated_device} = Devices.clear_penalty_box(device, user)
+    {:ok, updated_device} = Updates.clear_penalty_box(device, user)
 
     socket
     |> assign(:device, updated_device)
@@ -274,7 +275,7 @@ defmodule NervesHubWeb.Live.Devices.Show do
 
     authorized!(:"device:toggle-updates", current_scope)
 
-    {:ok, updated_device} = Devices.toggle_automatic_updates(device, user)
+    {:ok, updated_device} = Updates.toggle_automatic_updates(device, user)
 
     message = [
       "Firmware updates ",
