+++ lib/nerves_hub_web/live/devices/show.html.heex
@@ -94,7 +94,7 @@
         <%= cond do %>
           <% is_nil(@device.deployment_id) -> %>
             <span class="text-base-300 mr-1 ml-2 text-sm">N/A</span>
-          <% Devices.device_in_penalty_box?(@device) -> %>
+          <% Updates.device_in_penalty_box?(@device) -> %>
             <div class="relative z-20" id={"update-status-#{@device.id}"} phx-hook="ToolTip" data-placement="left">
               <svg class="stroke-warning mx-2 size-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none">
                 <path
