+++ test/nerves_hub/managed_deployments/distributed/orchestrator_test.exs
@@ -9,6 +9,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
   alias NervesHub.Devices.BulkActions
   alias NervesHub.Devices.Connections
   alias NervesHub.Devices.InflightUpdate
+  alias NervesHub.Devices.Updates
   alias NervesHub.Firmwares
   alias NervesHub.Firmwares.UpdateTool.Fwup
   alias NervesHub.Firmwares.Upload.File
@@ -230,13 +231,13 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     user: user,
     tmp_dir: tmp_dir
   } do
-    # `Devices.available_for_update` should be called:
+    # `Updates.available_for_update` should be called:
     # - once upon Orchestrator startup
     # - once when device1 is added to the deployment (but not online yet)
     # - once for when device1 comes online (out of date device comes online)
     # - once when device2 is added to the deployment (already up-to-date)
     # - and no more times after that
-    Devices
+    Updates
     |> expect(:available_for_update, 1, fn _deployment_group, _slots ->
       []
     end)
@@ -269,10 +270,10 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
         restart: :temporary
       })
 
-    allow(Devices, self(), pid)
+    allow(Updates, self(), pid)
 
     # only one device in this test isn't using the same firmware as the deployment group
-    # the `Devices.available_for_update/2` function should only be called once by device1
+    # the `Updates.available_for_update/2` function should only be called once by device1
 
     # assign device1 to the deployment group and mark it as 'connected'
     # this device will be told to update
@@ -299,7 +300,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     assert_receive %Broadcast{topic: ^device1_topic, event: "update"}, 1_000
 
     # device2 is already on the latest firmware, so when it comes online
-    # `Devices.available_for_update/2` won't be called and the device won't
+    # `Updates.available_for_update/2` won't be called and the device won't
     # be told to update
     device2_topic = "device:#{device2.id}"
     Phoenix.PubSub.subscribe(NervesHub.PubSub, device2_topic)
@@ -311,7 +312,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
 
     assert_receive %Broadcast{topic: ^deployment_group_topic, event: "device-added"}, 500
 
-    Mimic.reject(&Devices.available_for_update/2)
+    Mimic.reject(&Updates.available_for_update/2)
 
     {:ok, connection} = Connections.device_connecting(device2.org_id, device2.product_id, device2.id)
     :ok = Connections.device_connected(connection.id)
@@ -335,11 +336,11 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     tmp_dir: tmp_dir
   } do
     # An ugly set of expectations
-    # `Devices.available_for_update` should be called:
+    # `Updates.available_for_update` should be called:
     # - once upon Orchestrator startup
     # - and once when a bulk number of devices are added to the deployment group
     # - and no more times after that
-    Devices
+    Updates
     |> expect(:available_for_update, 1, fn _deployment_group, _slots ->
       []
     end)
@@ -370,13 +371,13 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
         restart: :temporary
       })
 
-    allow(Devices, self(), pid)
+    allow(Updates, self(), pid)
 
     BulkActions.move_many_to_deployment_group([device1.id, device2.id], deployment_group, user)
 
     assert_receive %Broadcast{topic: ^deployment_group_topic, event: "bulk-devices-added"}, 500
 
-    Mimic.reject(&Devices.available_for_update/2)
+    Mimic.reject(&Updates.available_for_update/2)
 
     # allows for db connections to finish and close
     :sys.get_state(pid)
@@ -432,7 +433,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     refute_receive %Broadcast{topic: ^device1_topic, event: "update"}, 1_000
 
     # we enable updates for the device
-    Devices.enable_updates(device1, user)
+    Updates.enable_updates(device1, user)
 
     # and then a device is told to schedule an update
     assert_receive %Broadcast{topic: ^device1_topic, event: "update"}, 1_000
@@ -465,11 +466,11 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     Phoenix.PubSub.subscribe(NervesHub.PubSub, deployment_topic)
 
     # An ugly set of expectations
-    # `Devices.available_for_update` should be called:
+    # `Updates.available_for_update` should be called:
     # - once upon Orchestrator startup
     # - once when a device is added to it
     # - and no more times after that
-    Devices
+    Updates
     |> expect(:available_for_update, 1, fn _deployment_group, _slots ->
       []
     end)
@@ -485,7 +486,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
         restart: :temporary
       })
 
-    allow(Devices, self(), pid)
+    allow(Updates, self(), pid)
 
     _device1 = Devices.update_deployment_group(device1, deployment_group)
 
@@ -563,7 +564,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
 
     expect(File, :upload_file, fn _, _ -> :ok end)
 
-    expect(Devices, :available_for_update, 1, fn _, _ -> [] end)
+    expect(Updates, :available_for_update, 1, fn _, _ -> [] end)
 
     {:ok, pid} =
       start_supervised(%{
@@ -572,7 +573,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
         restart: :temporary
       })
 
-    allow(Devices, self(), pid)
+    allow(Updates, self(), pid)
 
     :ok = Firmwares.generate_firmware_delta(delta, source_firmware, deployment_group.current_release.firmware)
 
@@ -606,9 +607,9 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
         restart: :temporary
       })
 
-    reject(&Devices.available_for_update/2)
+    reject(&Updates.available_for_update/2)
 
-    allow(Devices, self(), pid)
+    allow(Updates, self(), pid)
 
     {:ok, _} = ManagedDeployments.update_deployment_group(deployment_group, %{is_active: true}, user)
 
@@ -617,7 +618,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
 
   describe "trigger_update/1" do
     test "ignores updates when deployment_group is inactive", %{deployment_group: deployment_group} do
-      reject(&Devices.available_for_update/2)
+      reject(&Updates.available_for_update/2)
       reject(&Orchestrator.schedule_devices!/2)
 
       Orchestrator.trigger_update(%{deployment_group | is_active: false})
@@ -626,7 +627,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     test "skips scheduling firmware updates when deployment_group status is :preparing", %{
       deployment_group: deployment_group
     } do
-      reject(&Devices.available_for_update/2)
+      reject(&Updates.available_for_update/2)
       reject(&Orchestrator.schedule_devices!/2)
 
       Orchestrator.trigger_update(%{deployment_group | status: :preparing})
@@ -635,7 +636,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     test "skips scheduling firmware updates when deployment_group status is :deltas_failed", %{
       deployment_group: deployment_group
     } do
-      reject(&Devices.available_for_update/2)
+      reject(&Updates.available_for_update/2)
       reject(&Orchestrator.schedule_devices!/2)
 
       Orchestrator.trigger_update(%{deployment_group | status: :deltas_failed})
@@ -644,7 +645,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
     test "skips scheduling firmware updates when deployment_group status is :unknown_error", %{
       deployment_group: deployment_group
     } do
-      reject(&Devices.available_for_update/2)
+      reject(&Updates.available_for_update/2)
       reject(&Orchestrator.schedule_devices!/2)
 
       Orchestrator.trigger_update(%{deployment_group | status: :unknown_error})
@@ -871,7 +872,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
       :ok = Connections.device_connected(conn.id)
 
       # Should return empty list since priority queue is disabled
-      assert Devices.available_for_priority_update(deployment_group, 10) == []
+      assert Updates.available_for_priority_update(deployment_group, 10) == []
     end
 
     test "devices with version above threshold not eligible for priority queue", %{
@@ -909,7 +910,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
       {:ok, conn} = Connections.device_connecting(new_device.org_id, new_device.product_id, new_device.id)
       :ok = Connections.device_connected(conn.id)
       # new_device has version 1.5.0, threshold is 1.0.0
-      available = Devices.available_for_priority_update(deployment_group, 10)
+      available = Updates.available_for_priority_update(deployment_group, 10)
       refute Enum.any?(available, &(&1.id == new_device.id))
     end
 
@@ -961,7 +962,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
       :ok = Connections.device_connected(conn1.id)
       :ok = Connections.device_connected(conn2.id)
       # old devices have versions 0.9.0 and 0.8.0, threshold is 1.0.0
-      available = Devices.available_for_priority_update(deployment_group, 10)
+      available = Updates.available_for_priority_update(deployment_group, 10)
       device_ids = Enum.map(available, & &1.id)
 
       assert old_device1.id in device_ids
@@ -1017,7 +1018,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
 
       # The old `semver_match` raised casting "not-a-semver" to int[]; a NULL
       # `semver_sort_key` excludes the malformed device and the query succeeds.
-      device_ids = deployment_group |> Devices.available_for_priority_update(10) |> Enum.map(& &1.id)
+      device_ids = deployment_group |> Updates.available_for_priority_update(10) |> Enum.map(& &1.id)
 
       assert eligible.id in device_ids
       refute bad.id in device_ids
@@ -1182,7 +1183,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
       {:ok, conn} = Connections.device_connecting(device.org_id, device.product_id, device.id)
       :ok = Connections.device_connected(conn.id)
 
-      assert Devices.available_for_priority_update(deployment_group, 10) == []
+      assert Updates.available_for_priority_update(deployment_group, 10) == []
     end
 
     test "correctly handles semantic versioning with double-digit minor/patch versions", %{
@@ -1266,7 +1267,7 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
       :ok = Connections.device_connected(conn3.id)
       :ok = Connections.device_connected(conn4.id)
 
-      available = Devices.available_for_priority_update(deployment_group, 10)
+      available = Updates.available_for_priority_update(deployment_group, 10)
       device_ids = Enum.map(available, & &1.id)
 
       # With proper semantic versioning:
@@ -1456,9 +1457,9 @@ defmodule NervesHub.ManagedDeployments.Distributed.OrchestratorTest do
       # Give device1 enough failed update attempts to meet the failure threshold
       # The default device_failure_threshold is 3, so we'll add 3 attempts
       # This will cause it to fail can_device_update? and be skipped
-      :ok = to_device_info(old_device1) |> Devices.update_attempted()
-      :ok = to_device_info(old_device1) |> Devices.update_attempted()
-      :ok = to_device_info(old_device1) |> Devices.update_attempted()
+      :ok = to_device_info(old_device1) |> Updates.update_attempted()
+      :ok = to_device_info(old_device1) |> Updates.update_attempted()
+      :ok = to_device_info(old_device1) |> Updates.update_attempted()
 
       # Start orchestrator
       {:ok, pid} =
