+++ test/test_helper.exs
@@ -4,6 +4,7 @@ Mimic.copy(ProcessHub.Future)
 Mimic.copy(ProcessHub.StartResult)
 Mimic.copy(ProcessHub.Future)
 Mimic.copy(NervesHub.Devices)
+Mimic.copy(NervesHub.Devices.Updates)
 Mimic.copy(NervesHub.ManagedDeployments.Distributed.Orchestrator)
 Mimic.copy(NervesHub.Firmwares)
 Mimic.copy(NervesHub.Firmwares.UpdateTool.Fwup)
